Make launchd compile 64-bit. It still doesn't work. What remains is malloc()/free() problems (in launchd_unix_ipc.c???).
git-svn-id: http://svn.macosforge.org/repository/launchd/trunk@23519 e92bca22-270c-0410-9cea-e3f1106b6a1c

+++ launchd/src/launchd_runtime.h
@@ -163,7 +163,7 @@ typedef enum {
 /* All of these log the return address as "arg4" */
 INTERNAL_ABI void runtime_ktrace1(runtime_ktrace_code_t code);
 INTERNAL_ABI void runtime_ktrace0(runtime_ktrace_code_t code);
-INTERNAL_ABI void runtime_ktrace(runtime_ktrace_code_t code, int a, int b, int c);
+INTERNAL_ABI void runtime_ktrace(runtime_ktrace_code_t code, long a, long b, long c);
 
 
 #define LOG_APPLEONLY 0x4141504c /* AAPL in hex */
